update: refactor useFullTextSearch hook
